week6/offers/requirements.txt
appdirs==1.4.3
coverage==4.3.4
decorator==4.0.11
Django==1.11
django-debug-toolbar==1.7
django-money==0.10.2
djangorestframework==3.6.2
djangorestframework-jwt==1.10.0
factory-boy==2.8.1
Faker==0.7.10
ipdb==0.10.2
ipython==5.3.0
ipython-genutils==0.2.0
olefile==0.44
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
prompt-toolkit==1.0.14
psycopg2==2.7.1
ptyprocess==0.5.1
py-moneyed==0.6.0
Pygments==2.2.0
PyJWT==1.5.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.2.3
traitlets==4.3.2
wcwidth==0.1.7
